+++ projectroles/forms.py
@@ -33,6 +33,7 @@ from projectroles.utils import get_display_name, build_secret
 from projectroles.app_settings import AppSettingAPI
 
 
+app_settings = AppSettingAPI()
 User = auth.get_user_model()
 
 
@@ -46,6 +47,7 @@ PROJECT_TYPE_PROJECT = SODAR_CONSTANTS['PROJECT_TYPE_PROJECT']
 SITE_MODE_SOURCE = SODAR_CONSTANTS['SITE_MODE_SOURCE']
 SITE_MODE_TARGET = SODAR_CONSTANTS['SITE_MODE_TARGET']
 APP_SETTING_SCOPE_PROJECT = SODAR_CONSTANTS['APP_SETTING_SCOPE_PROJECT']
+APP_SETTING_SCOPE_SITE = SODAR_CONSTANTS['APP_SETTING_SCOPE_SITE']
 APP_SETTING_TYPE_BOOLEAN = SODAR_CONSTANTS['APP_SETTING_TYPE_BOOLEAN']
 APP_SETTING_TYPE_INTEGER = SODAR_CONSTANTS['APP_SETTING_TYPE_INTEGER']
 APP_SETTING_TYPE_JSON = SODAR_CONSTANTS['APP_SETTING_TYPE_JSON']
@@ -259,7 +261,7 @@ class SODARUserChoiceField(forms.ModelChoiceField):
             limit_choices_to=None,
             widget=widget,
             *args,
-            **kwargs
+            **kwargs,
         )
 
 
@@ -418,7 +420,7 @@ class ProjectForm(SODARModelForm):
                     )
                     for value in values
                 ],
-                **setting_kwargs
+                **setting_kwargs,
             )
         elif s_def.options and callable(s_def.options) and not self.instance.pk:
             values = s_def.options(project=None)
@@ -431,7 +433,7 @@ class ProjectForm(SODARModelForm):
                     )
                     for value in values
                 ],
-                **setting_kwargs
+                **setting_kwargs,
             )
         elif s_def.options:
             self.fields[s_field] = forms.ChoiceField(
@@ -443,7 +445,7 @@ class ProjectForm(SODARModelForm):
                     )
                     for option in s_def.options
                 ],
-                **setting_kwargs
+                **setting_kwargs,
             )
         # Other types
         elif s_def.type == APP_SETTING_TYPE_STRING:
@@ -468,10 +470,9 @@ class ProjectForm(SODARModelForm):
             )
 
         # Add optional attributes from plugin (#404)
-        # NOTE: Experimental! Use at your own risk!
         self.fields[s_field].widget.attrs.update(s_widget_attrs)
         # Set initial value
-        value = self.app_settings.get(
+        value = app_settings.get(
             plugin_name=plugin_name,
             setting_name=s_def.name,
             project=self.instance if self.instance.pk else None,
@@ -513,20 +514,19 @@ class ProjectForm(SODARModelForm):
         # Show unmodifiable settings to superusers
         if not self.current_user.is_superuser:
             self.p_kwargs['user_modifiable'] = True
-        self.app_settings = AppSettingAPI()
         self.app_plugins = sorted(get_active_plugins(), key=lambda x: x.name)
         for plugin in self.app_plugins + [None]:  # Projectroles has no plugin
             if plugin:
                 plugin_name = plugin.name
-                s_defs = self.app_settings.get_definitions(
+                s_defs = app_settings.get_definitions(
                     APP_SETTING_SCOPE_PROJECT, plugin=plugin, **self.p_kwargs
                 )
             else:
                 plugin_name = APP_NAME
-                s_defs = self.app_settings.get_definitions(
+                s_defs = app_settings.get_definitions(
                     APP_SETTING_SCOPE_PROJECT,
                     plugin_name=plugin_name,
-                    **self.p_kwargs
+                    **self.p_kwargs,
                 )
             for s_def in s_defs.values():
                 s_field = 'settings.{}.{}'.format(plugin_name, s_def.name)
@@ -795,7 +795,7 @@ class ProjectForm(SODARModelForm):
         cleaned_data, errors = self._validate_app_settings(
             self.cleaned_data,
             self.app_plugins,
-            self.app_settings,
+            app_settings,
             self.p_kwargs,
             self.instance,
         )
@@ -1186,6 +1186,177 @@ class ProjectInviteForm(SODARModelForm):
         obj.secret = build_secret()
         obj.save()
         return obj
+
+
+# Site app settings form -------------------------------------------------------
+
+
+class SiteAppSettingsForm(SODARForm):
+    """Form for managing site app settings"""
+
+    # TODO: Unify repeated code with ProjectForm and UserAppSettingsForm
+    def _set_app_setting_field(self, plugin_name, s_field, s_def):
+        """
+        Set site app setting field, widget and value.
+
+        :param plugin_name: App plugin name
+        :param s_field: Form field name
+        :param s_def: PluginAppSettingDef object
+        """
+        s_widget_attrs = s_def.widget_attrs
+        if s_def.placeholder is not None:
+            s_widget_attrs['placeholder'] = s_def.placeholder
+        setting_kwargs = {
+            'required': False,
+            'label': s_def.label or '{}.{}'.format(plugin_name, s_def.name),
+            'help_text': s_def.description,
+        }
+        # Disable global settings if on target site
+        if (
+            s_def.global_edit
+            and settings.PROJECTROLES_SITE_MODE == SITE_MODE_TARGET
+        ):
+            setting_kwargs['label'] += ' ' + SETTING_DISABLE_LABEL
+            setting_kwargs['help_text'] += ' ' + SETTING_SOURCE_ONLY_MSG
+            setting_kwargs['disabled'] = True
+
+        if s_def.options and callable(s_def.options):
+            self.fields[s_field] = forms.ChoiceField(
+                choices=[
+                    (
+                        (str(value[0]), str(value[1]))
+                        if isinstance(value, tuple)
+                        else (str(value), str(value))
+                    )
+                    for value in s_def.options()
+                ],
+                **setting_kwargs,
+            )
+        elif s_def.options:
+            self.fields[s_field] = forms.ChoiceField(
+                choices=[
+                    (
+                        (int(option), int(option))
+                        if s_def.type == APP_SETTING_TYPE_INTEGER
+                        else (option, option)
+                    )
+                    for option in s_def.options
+                ],
+                **setting_kwargs,
+            )
+        elif s_def.type == APP_SETTING_TYPE_STRING:
+            self.fields[s_field] = forms.CharField(
+                widget=forms.TextInput(attrs=s_widget_attrs),
+                **setting_kwargs,
+            )
+        elif s_def.type == APP_SETTING_TYPE_INTEGER:
+            self.fields[s_field] = forms.IntegerField(
+                widget=forms.NumberInput(attrs=s_widget_attrs),
+                **setting_kwargs,
+            )
+        elif s_def.type == APP_SETTING_TYPE_BOOLEAN:
+            self.fields[s_field] = forms.BooleanField(**setting_kwargs)
+        elif s_def.type == APP_SETTING_TYPE_JSON:
+            # NOTE: Attrs MUST be supplied here (#404)
+            if 'class' in s_widget_attrs:
+                s_widget_attrs['class'] += ' sodar-json-input'
+            else:
+                s_widget_attrs['class'] = 'sodar-json-input'
+            self.fields[s_field] = forms.CharField(
+                widget=forms.Textarea(attrs=s_widget_attrs),
+                **setting_kwargs,
+            )
+
+        # Modify initial value and attributes
+        self.fields[s_field].widget.attrs.update(s_widget_attrs)
+        value = app_settings.get(
+            plugin_name=plugin_name, setting_name=s_def.name
+        )
+        if s_def.type == APP_SETTING_TYPE_JSON:
+            value = json.dumps(value)
+        self.initial[s_field] = value
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        # Add settings fields
+        self.project_plugins = get_active_plugins(plugin_type='project_app')
+        self.site_plugins = get_active_plugins(plugin_type='site_app')
+        self.app_plugins = self.project_plugins + self.site_plugins
+
+        for plugin in self.app_plugins + [None]:
+            if plugin:
+                plugin_name = plugin.name
+                s_defs = app_settings.get_definitions(
+                    APP_SETTING_SCOPE_SITE, plugin=plugin, user_modifiable=True
+                )
+            else:
+                plugin_name = 'projectroles'
+                s_defs = app_settings.get_definitions(
+                    APP_SETTING_SCOPE_SITE,
+                    plugin_name=plugin_name,
+                    user_modifiable=True,
+                )
+            for s_def in s_defs.values():
+                s_field = 'settings.{}.{}'.format(plugin_name, s_def.name)
+                self._set_app_setting_field(plugin_name, s_field, s_def)
+                self.fields[s_field].label = self.get_app_setting_label(
+                    plugin, self.fields[s_field].label
+                )
+
+    def clean(self):
+        for plugin in self.app_plugins + [None]:
+            p_kwargs = {'user_modifiable': True}
+            if plugin:
+                p_name = plugin.name
+                p_kwargs['plugin'] = plugin
+            else:
+                p_name = 'projectroles'
+                p_kwargs['plugin_name'] = p_name
+            s_defs = app_settings.get_definitions(
+                APP_SETTING_SCOPE_SITE, **p_kwargs
+            )
+            p_settings = {}
+
+            for s_def in s_defs.values():
+                s_field = '.'.join(['settings', p_name, s_def.name])
+                p_settings[s_def.name] = self.cleaned_data.get(s_field)
+
+                if s_def.type == APP_SETTING_TYPE_JSON:
+                    if not self.cleaned_data.get(s_field):
+                        self.cleaned_data[s_field] = '{}'
+                    try:
+                        self.cleaned_data[s_field] = json.loads(
+                            self.cleaned_data.get(s_field)
+                        )
+                    except json.JSONDecodeError as err:
+                        self.add_error(s_field, 'Invalid JSON\n' + str(err))
+                elif s_def.type == APP_SETTING_TYPE_INTEGER:
+                    # Convert integers from select fields
+                    self.cleaned_data[s_field] = int(self.cleaned_data[s_field])
+
+                if not app_settings.validate(
+                    setting_type=s_def.type,
+                    setting_value=self.cleaned_data.get(s_field),
+                    setting_options=s_def.options,
+                ):
+                    self.add_error(s_field, 'Invalid value')
+
+            # Custom plugin validation for app settings
+            if plugin and hasattr(plugin, 'validate_form_app_settings'):
+                try:
+                    p_errors = plugin.validate_form_app_settings(p_settings)
+                    if p_errors:
+                        for field, error in p_errors.items():
+                            f_name = '.'.join(['settings', p_name, field])
+                            self.add_error(f_name, error)
+                except Exception as ex:
+                    self.add_error(
+                        None,
+                        SETTING_CUSTOM_VALIDATE_MSG.format(
+                            plugin=p_name, exception=ex
+                        ),
+                    )
+        return self.cleaned_data
 
 
 # RemoteSite form --------------------------------------------------------------